+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Charge Pads now accept four rechargeable items in their UI while continuing to charge carried and worn items when stood on. ([Issue 2123](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2123))
 * Scrap Boxes now give a random reward when used. Reward selection is based on [PR 56](https://github.com/FTBTeam/FTB-Industrial-Contraptions/pull/56) by @jshipley. ([Issue 2124](https://github.com/FTBTeam/FTB-Mods-Issues/issues/2124))
 
+### Added
+
+* Items with the `ftbic:loot_box` data component can open a chosen loot table. Crouching opens the whole stack at once. Based on [PR 56](https://github.com/FTBTeam/FTB-Industrial-Contraptions/pull/56) by @jshipley.
+
+### Changed
+
+* Canned Food now uses the standard item remainder behavior to return its empty can after eating.
+
 ## [26.1.2.10]
 
 ### Fixed